feat(user): move anmelde-versuche to benutzer, refine userguide pdf/dashboard
- Extend userguide PDF page-break CSS: keep an intro paragraph glued to its
  screenshot so a page never starts with a bare image, keep bullet/numbered
  lists intact on one page, and add a thin border around every screenshot.
- Move the Anmelde-Versuche (login attempts) admin screen from the settings
  module to the user module: route benutzer/anmelde-versuche, backend
  endpoints on UserController under USER_MANAGEMENT instead of SETTINGS.
- Seed permanent login-attempts testdata (previously empty) and update the
  userguide screenshot/chapter accordingly.
- Fix the dashboard "Erfasste Routen (Details)" placeholder dash to match the
  text size used by every other dashboard panel, while keeping actual route
  names at a smaller size so long lists still fit.

Closes #3017
