+++ frontend/src/main/webapp/src/app/api/settings-api.service.spec.ts
@@ -1,9 +1,7 @@
 import {HttpTestingController, provideHttpClientTesting} from '@angular/common/http/testing';
 import {TestBed} from '@angular/core/testing';
-import {UserApiService} from './user-api.service';
 import {provideHttpClient, withXhr} from '@angular/common/http';
 import {
-  LoginAttemptItem,
   MailRecipients,
   MailTypeEnum,
   RecipientTypeEnum,
@@ -11,7 +9,6 @@ import {
   StaticValueItem,
   StaticValueTypeEnum
 } from './settings-api.service';
-import {PagedResponse} from '../common/api/paged-response';
 
 describe('SettingsApiService', () => {
   let httpMock: HttpTestingController;
@@ -21,8 +18,7 @@ describe('SettingsApiService', () => {
     TestBed.configureTestingModule({
       providers: [
         provideHttpClient(withXhr()),
-        provideHttpClientTesting(),
-        UserApiService
+        provideHttpClientTesting()
       ]
     });
 
@@ -98,46 +94,6 @@ describe('SettingsApiService', () => {
     httpMock.verify();
 
     expect(req.request.body).toEqual(testData);
-  });
-
-  it('get login attempts', () => {
-    const testResponse: PagedResponse<LoginAttemptItem> = {
-      items: [
-        {id: 1, username: 'user1', failureCount: 3, lastFailureAt: '2026-01-01T10:00:00', lockedUntil: '2026-01-01T10:15:00'},
-        {id: 2, username: 'user2', failureCount: 1, lastFailureAt: '2026-01-01T09:00:00', lockedUntil: null}
-      ],
-      totalCount: 2,
-      currentPage: 1,
-      totalPages: 1,
-      pageSize: 10
-    };
-
-    apiService.getLoginAttempts().subscribe((data: PagedResponse<LoginAttemptItem>) => {
-      expect(data).toEqual(testResponse);
-    });
-
-    const req = httpMock.expectOne({method: 'GET', url: '/settings/login-attempts'});
-    req.flush(testResponse);
-    httpMock.verify();
-  });
-
-  it('get login attempts with page and pageSize', () => {
-    const page = 2;
-    const pageSize = 25;
-
-    apiService.getLoginAttempts(page, pageSize).subscribe();
-
-    const req = httpMock.expectOne({method: 'GET', url: `/settings/login-attempts?page=${page}&pageSize=${pageSize}`});
-    req.flush({items: []});
-    httpMock.verify();
-  });
-
-  it('delete login attempt', () => {
-    apiService.deleteLoginAttempt(1).subscribe();
-
-    const req = httpMock.expectOne({method: 'DELETE', url: '/settings/login-attempts/1'});
-    req.flush(null);
-    httpMock.verify();
   });
 
 });
