+++ frontend/src/main/webapp/src/app/modules/dashboard/components/recorded-route-names/recorded-route-names.component.html
@@ -8,7 +8,7 @@
     @if (recordedRouteNames() && recordedRouteNames()!.length > 0) {
       <div testid="recorded-route-names" class="text-white text-3xl">{{ recordedRouteNames()!.join(', ') }}</div>
     } @else {
-      <div testid="recorded-route-names" class="text-white text-3xl">-</div>
+      <div testid="recorded-route-names" class="text-white text-5xl">-</div>
     }
   </mat-card-content>
 </mat-card>
