+++ frontend/src/main/webapp/src/app/modules/user/views/login-attempts/dialogs/delete-login-attempt-dialog.component.ts
